Combined
I have sorted all the files into the appropriate folders. I have connected all the code up and tested it. I removed the jquery file as we can just use the online link. The google sign has not been implemented into this code as of yet. The Sign in and Sign up work and it checks that the passwords match before creating the account.